delete user with admin works now

+++ api/admin.py
@@ -12,8 +12,8 @@ async def delete_all_data_from_tables(authorization: str = Header(None)):
     try:
         authenticated = hranhsjwt.check_user_role(authorization)
         if authenticated:
-            authorized = hranhsjwt.authenticate_user(authorization)
-            if authorized.role == "admin":
+            admin_only = hranhsjwt.restrict_access(authorization)
+            if admin_only:
                 hracrud.delete_all_data_from_tables(User.USERSTABLENAME)
                 hracrud.delete_all_data_from_tables(Vendor.VENDORTABLENAME)
                 hracrud.delete_all_data_from_tables(MedicineAsset.MEDICINEASSETSTABLENAME)
@@ -27,41 +27,64 @@ async def delete_all_data_from_tables(authorization: str = Header(None)):
         print(type(ex),ex)
         return {"error":f"{type(ex)},{ex}"}
 @router.delete("/delete_vendor/{vendor_id}")
-async def delete_vendor(vendor_id:Optional[str] = None,authorization: str = Header(None)):
+async def delete_vendor(vendor_id:str,authorization: str = Header(None)):
     try:
         authenticated = hranhsjwt.check_user_role(authorization)
         if authenticated:
-            if vendor_id:   
+            admin_only = hranhsjwt.restrict_access(authorization)
+            if admin_only:
                 condition = f"{Vendor.get_field_name('vendor_id')} = '{vendor_id}'" 
+                asset_exists = hracrud.check_exists(("*"),Vendor.VENDORTABLENAME,condition=condition)
+                if asset_exists:
+                    hracrud.delete_data(Vendor.VENDORTABLENAME,condition=condition)
+                    hracrud.delete_data(MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
+                    return {"message":"Vendor was deleted."}
+                else:
+                    return {"error":"No assets found."} 
             else:
-                return {"error":"Please provide vendor_id or vendor."}
-            asset_exists = hracrud.check_exists(("*"),Vendor.VENDORTABLENAME,condition=condition)
-            if asset_exists:
-                hracrud.delete_data(Vendor.VENDORTABLENAME,condition=condition)
-                hracrud.delete_data(MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
-                return {"message":"Vendor was deleted."}
-            else:
-                return {"error":"No assets found."} 
+                return {"error":"User is not authorized."}
         else:
             return {"error":"User does not exist or is not authorized."}
     except Exception as ex:
         print(type(ex),ex)
         return {"error":f"{type(ex)},{ex}"}
 @router.delete("/delete_medicine_asset/{medicine_id}")
-async def delete_medicine_asset(medicine_id:Optional[str] = None,authorization: str = Header(None)):
+async def delete_medicine_asset(medicine_id:str,authorization: str = Header(None)):
     try:
         authenticated = hranhsjwt.check_user_role(authorization)
         if authenticated:
-            if medicine_id:   
+            admin_only = hranhsjwt.restrict_access(authorization)
+            if admin_only:
                 condition = f"{MedicineAsset.get_field_name('medicine_id')} = '{medicine_id}'" 
+                asset_exists = hracrud.check_exists(("*"),MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
+                if asset_exists:
+                    hracrud.delete_data(MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
+                    return {"message":"Medicine Asset was deleted."}
+                else:
+                    return {"error":"No assets found."} 
             else:
-                return {"error":"Please provide medicine_id or medicine_asset."}
-            asset_exists = hracrud.check_exists(("*"),MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
-            if asset_exists:
-                hracrud.delete_data(MedicineAsset.MEDICINEASSETSTABLENAME,condition=condition)
-                return {"message":"Medicine Asset was deleted."}
+                return {"error":"User is not authorized."}
+        else:
+            return {"error":"User does not exist or is not authorized."}
+    except Exception as ex:
+        print(type(ex),ex)
+        return {"error":f"{type(ex)},{ex}"}
+@router.delete("/delete_user/{user_id}")
+async def delete_user(user_id:str,authorization: str = Header(None)):
+    try:
+        authenticated = hranhsjwt.check_user_role(authorization)
+        if authenticated:
+            admin_only = hranhsjwt.restrict_access(authorization)
+            if admin_only:
+                condition = f"{User.get_field_name('user_id')} = '{user_id}'" 
+                user_exists = hracrud.check_exists(("*"),User.USERSTABLENAME,condition=condition)
+                if user_exists:
+                    hracrud.delete_data(User.USERSTABLENAME,condition=condition)
+                    return {"message":"User was deleted."}
+                else:
+                    return {"error":"No assets found."} 
             else:
-                return {"error":"No assets found."} 
+                return {"error":"User is not authorized."}
         else:
             return {"error":"User does not exist or is not authorized."}
     except Exception as ex:
